Do not delete ZIM of extra.e2e.test.ts as well when requested for

+++ test/e2e/extra.e2e.test.ts
@@ -59,7 +59,9 @@ describe('Extra', () => {
 
       // Scraped customMainPage
       // TODO: clear test dir
-      rimraf.sync(testId)
+      if (!process.env.KEEP_ZIMS) {
+        rimraf.sync(testId)
+      }
 
       const redisScan = await execa('redis-cli --scan', { shell: true })
       // Redis has been cleared
